pyproject.toml
[tool.poetry]
name = "ibis-substrait"
version = "1.0.0"
packages = [{ include = "ibis_substrait" }]
homepage = "https://github.com/ibis-project/ibis-substrait"
repository = "https://github.com/ibis-project/ibis-substrait"
description = "Subtrait compiler for ibis"
authors = ["Ibis Contributors"]
maintainers = ["Ibis Contributors"]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
ibis-framework = "^2.0.0"
inflection = "^0.5.1"
cached-property = "^1.5.2"
pyarrow = "^6.0.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^4.0.0"
isort = "^5.9.3"
mypy = "^0.931"
protoletariat = "^0.7.0"
pydocstyle = "^6.1.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-mock = "^3.6.1"
pytest-randomly = "^3.10.1"
pyupgrade = "^2.26.0"
types-protobuf = "^3.18.1"
mypy-protobuf = "^3.0.0"
pytz = "^2021.3"
types-pytz = "^2021.3.4"
ptpython = "^3.0.20"

[tool.pytest.ini_options]
xfail_strict = true
addopts = [
  "--ignore=site-packages",
  "--ignore=dist-packages",
  "--ignore=.direnv",
  "--strict-markers",
  "--doctest-modules",
]
markers = ["no_decompile"]
norecursedirs = ["site-packages", "dist-packages", ".direnv"]

[tool.black]
line_length = 88
exclude = "ibis_substrait/proto"

[tool.isort]
profile = "black"
skip_glob = ["ibis_substrait/proto/**/*.py"]

[tool.mypy]
exclude = "ibis_substrait/proto/.+\\.py|tests/.+\\.py"
ignore_missing_imports = true
# untyped things
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# Any handling
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = false
disallow_subclassing_any = false
# None/Optional handling
no_implicit_optional = true
# show more context on failure
show_error_context = true
show_error_codes = true

[tool.pydocstyle]
inherit = false
convention = "numpy"
match_dir = "ibis_substrait"
add_ignore = ["D100", "D101", "D102", "D103", "D104", "D105"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
